连接centos增加一组数据库密码，完善警告

+++ rdata.cpp
@@ -50,7 +50,7 @@ bool Rdata::stoxml(int op, char *buffer)
     }
     s += "</data>";
 	memcpy(buffer, s.c_str(), s.size() + 1);
-    len = strlen(buffer);
+    len = (int)strlen(buffer);
     return true;
 }
